Final update
Added node files, in the folder load_test, that can be used to test the load node feature. Added node files to the output folder, which were results of the node_generate feature. Improved the load feature in the parser file in order to get it to work more efficiently. Fixed some functions in the nodecaster file.